compiler erro fixes

+++ src/Image.cpp
@@ -4,8 +4,8 @@
 #include "stb_image_write.h"
 #include <spdlog/spdlog.h>
 
-void writePNG(const char* filename, const int w, const int h, const int c, const std::vector<uint8_t>& image)
+void writePNG(const char* filename, const uint32_t w, const uint32_t h, const uint32_t c, const std::vector<uint8_t>& image)
 {
 	spdlog::critical("Writing to file: {}", filename);
-	stbi_write_png(filename, w, h, c, image.data(), w * c);
+	stbi_write_png(filename, static_cast<int>(w), static_cast<int>(h), static_cast<int>(c), image.data(), static_cast<int>(w * c));
 }
